Implemented fake hash lookup solution
This is only a partial implemention, I tested and you can add multiple
args to a string and do combined lookup on more than one variable, you
just need to manage the elements and ordering of them when you make each
one. This works ok but would get crazy if you needed to work with a
large number of items as the fat finger potential is exponential.